requirements.txt
apturl==0.5.2
asgiref==3.4.1
asn1crypto==0.24.0
astroid==2.4.2
autopep8==1.5.5
bcrypt==3.1.7
blinker==1.4
Brlapi==0.7.0
Brotli==1.0.7
certifi==2019.11.28
chardet==3.0.4
Click==7.0
colorama==0.4.3
command-not-found==0.3
cryptography==2.8
cupshelpers==1.0
dbus-python==1.2.16
defer==1.0.6
distro==1.4.0
distro-info===0.23ubuntu1
dj-database-url==0.5.0
Django==3.2.7
django-cors-headers==3.3.0
django-filter==2.2.0
django-jsonfield==1.4.0
djangorestframework==3.11.0
duplicity==0.8.12.0
entrypoints==0.3
fasteners==0.14.1
future==0.18.2
gpg===1.13.1-unknown
gunicorn==20.0.4
httplib2==0.14.0
idna==2.8
isort==4.3.21
keyring==18.0.1
keyrings.alt==3.4.0
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lazy-object-proxy==1.4.3
lockfile==0.12.2
louis==3.12.0
macaroonbakery==1.3.1
Mako==1.1.0
Markdown==3.2.2
MarkupSafe==1.1.0
mccabe==0.6.1
monotonic==1.5
netifaces==0.10.4
oauth==1.0.1
oauthlib==3.1.0
olefile==0.46
paramiko==2.6.0
pexpect==4.6.0
Pillow==7.0.0
protobuf==3.6.1
psycopg2==2.9.1
pycairo==1.16.2
pycodestyle==2.7.0
pycrypto==2.6.1
pycups==1.9.73
Pygments==2.3.1
PyGObject==3.36.0
PyJWT==1.7.1
pylint==2.5.3
pymacaroons==0.13.0
PyNaCl==1.3.0
PyQt5==5.14.1
pyRFC3339==1.1
PySocks==1.6.8
python-apt==2.0.0+ubuntu0.20.4.6
python-dateutil==2.7.3
python-debian===0.1.36ubuntu1
pytz==2019.3
pyxdg==0.26
PyYAML==5.3.1
reportlab==3.5.34
requests==2.22.0
requests-unixsocket==0.2.0
SecretStorage==2.3.1
simplejson==3.16.0
sip==4.19.21
six==1.14.0
sqlparse==0.2.4
system-service==0.3
systemd-python==234
toml==0.10.1
torbrowser-launcher==0.3.2
ubuntu-advantage-tools==27.2
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.8
usb-creator==0.3.7
wadllib==1.3.3
whitenoise==5.1.0
wrapt==1.12.1
xkit==0.0.0
zope.interface==4.7.1
